+++ src/components/candidateContent/candidateFeedbacks/CandidateFeedbacks.sass
@@ -1,16 +1,5 @@
-.feedbackContainer
+.feedbacksContainer
   height: 100%
   display: flex
   flex-direction: column
-
-.feedbackItem
-  display: flex
-  flex-direction: column
-  margin-bottom: 30px
-
-.titleSection
-  height: 55px
-  display: flex
-  flex-direction: row
-  align-items: center
-  justify-content: space-between
+  padding-right: 20px
