+++ src/components/candidateContent/candidateFeedbacksItem/CandidateFeedbacksItem.sass
@@ -1,0 +1,11 @@
+.feedbackItem
+  display: flex
+  flex-direction: column
+  margin-bottom: 30px
+
+.titleSection
+  height: 55px
+  display: flex
+  flex-direction: row
+  align-items: center
+  justify-content: space-between
