+++ src/components/candidateContent/candidateFeedbacksItem/index.js
@@ -1,0 +1,1 @@
+export { default as CandidateFeedbacksItem } from './CandidateFeedbacksItem';
