+++ src/components/candidateContent/index.js
@@ -1,2 +1,4 @@
 export { CandidateInfo } from './candidateInfo';
 export { CandidateFeedbacks } from './candidateFeedbacks';
+export { CandidateFeedbacksItem } from './candidateFeedbacksItem';
+export { StarRating } from './starRating';
