+++ src/components/candidateContent/starRating/index.js
@@ -1,0 +1,1 @@
+export { default as StarRating } from './StarRating';
